+++ resources/Materials/TestSuite/stdlib/definition/surfacematerial_definition.mtlx
@@ -1,20 +1,31 @@
 <?xml version="1.0"?>
 <materialx version="1.37">
-  <nodedef name="ND_mymaterial" node="mymaterial" nodegroup="material">
+  <!-- Nodegraph definition with material node -->
+  <nodedef name="ND_mymaterial" node="mymaterial" nodegroup="material" target="mytarget" version="1.3" isdefaultversion="true">
     <output name="out" type="material"/>
   </nodedef>
-  <nodegraph name="NG_mymaterial" nodedef="ND_mymaterial">
+  <nodegraph name="NG_mymaterial" nodedef="ND_mymaterial" target="mytarget" version="1.3">
     <surfacematerial name="surfacematerial1" type="material">
       <input name="surfaceshader" type="surfaceshader" nodename="standard_surface"/>
     </surfacematerial>
     <standard_surface name="standard_surface" type="surfaceshader" />
     <output name="out" type="material" nodename="surfacematerial1" />
-  </nodegraph>
+   </nodegraph>
 
-   <nodedef name="ND_myshader" node="myshader" nodegroup="shader">
+   <!-- Nodegraph with material node -->
+   <nodegraph name="NG_mymaterial_instance">
+      <surfacematerial name="surfacematerial2" type="material">
+         <input name="surfaceshader" type="surfaceshader" nodename="standard_surface"/>
+      </surfacematerial>
+      <standard_surface name="standard_surface" type="surfaceshader" />
+      <output name="out" type="material" nodename="surfacematerial2" />
+   </nodegraph>
+
+   <!-- Nodegraph definition with shader node -->
+   <nodedef name="ND_myshader" node="myshader" nodegroup="shader" target="mytarget" version="1.3" isdefaultversion="true">
       <output name="surfaceshader" type="surfaceshader"/>
    </nodedef>
-   <nodegraph name="NG_myshader" nodedef="ND_myshader">
+   <nodegraph name="NG_myshader" nodedef="ND_myshader" target="mytarget" version="1.3">
       <standard_surface name="my_surface_shader" type="surfaceshader" />
       <output name="surfaceshader" type="surfaceshader" nodename="my_surface_shader" />
    </nodegraph>
