+++ source/MaterialXGenOgsXml/OgsFragment.cpp
@@ -1,5 +1,6 @@
 #include "OgsFragment.h"
 
+#include <MaterialXCore/MaterialNode.h>
 #include <MaterialXFormat/XmlIo.h>
 #include <MaterialXGenShader/Util.h>
 #include <MaterialXGenShader/DefaultColorManagementSystem.h>
